# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Admin API
  description: 工作流平台管理后台 REST API
  version: 1.0.0
  contact:
    name: Workflow Platform Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
...
paths:
...
  /api/admin/executions:
    get:
      operationId: listExecutions
      tags: [Executions]
      summary: 查询执行历史列表
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: workflowId
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 执行历史列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseExecutionRecord'
  /api/admin/executions/{executionId}:
    get:
      operationId: getExecution
      tags: [Executions]
      summary: 查看执行详情
      parameters:
        - $ref: '#/components/parameters/ExecutionIdPath'
      responses:
        '200':
          description: 执行详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
  /api/admin/executions/{executionId}/rerun:
    post:
      operationId: rerunExecution
      tags: [Executions]
      summary: 从指定节点重放执行
      parameters:
        - $ref: '#/components/parameters/ExecutionIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RerunExecutionRequest'
      responses:
        '200':
          description: 重放结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
  /api/admin/executions/{executionId}/signal:
    post:
      operationId: sendSignal
      tags: [Executions]
      summary: 向运行中的工作流发送信号（恢复等待节点）
      parameters:
        - $ref: '#/components/parameters/ExecutionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signalName]
              properties:
                signalName:
                  type: string
                  description: 信号名称
                data:
                  description: 信号携带的数据（任意类型）
      responses:
        '200':
          description: 信号已发送
          content:
            application/json:
              schema:
                type: object
                properties:
                  executionId:
                    type: string
                  signalName:
                    type: string
                  status:
                    type: string
                    enum: [DELIVERED, BUFFERED]
        '400':
          description: 信号未被接受（工作流可能未在运行）
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: 页码（从 0 开始）
      schema:
        type: integer
        default: 0
    PageSize:
      name: pageSize
      in: query
      description: 每页条数
      schema:
        type: integer
        default: 20
...
    ExecutionIdPath:
      name: executionId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    TraceDetail:
      type: object
      properties:
        executionId:
          type: string
        workflowName:
          type: string
        totalDurationMs:
          type: integer
          format: int64
        status:
          type: string
        traces:
          type: array
          items:
            $ref: '#/components/schemas/TraceNode'
    TraceNode:
      type: object
      properties:
        nodeId:
          type: string
        nodeName:
          type: string
        durationMs:
          type: integer
          format: int64
        status:
          type: string
        inputs:
          type: object
          additionalProperties: true
          description: 工作流各节点输入参数（key=nodeId）
        output:
          type: object
          additionalProperties: true
          description: 工作流各节点输入参数（key=nodeId）
        error:
          type: string
        sql:
          type: string

    # ============ User 相关 ============
...
    RerunExecutionRequest:
      type: object
      properties:
        nodeId:
          type: string

    # ============ 装饰器 ============
...
    ExecutionRecord:
      type: object
      properties:
        executionId:
          type: string
        workflowId:
          type: string
        workflowName:
          type: string
        status:
          type: string
        totalDurationMs:
          type: integer
          format: int64
        startTime:
          type: integer
          format: int64
          description: Unix 毫秒时间戳
        endTime:
          type: integer
          format: int64
          description: Unix 毫秒时间戳
    PageResponseExecutionRecord:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionRecord'
        total:
          type: integer
          format: int64
        page:
          type: integer
        pageSize:
          type: integer

    # ============ 监控 ============
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT